modularization of scripts
