+++ static/modules/layout.js
@@ -1,0 +1,24 @@
+function adjustIframeSizes() {
+  const iframeContainer = document.getElementById('iframeContainer');
+  const iframeBoxes = document.querySelectorAll('.iframe-box');
+  const channelCount = iframeBoxes.length;
+
+  let columns, rows;
+  if (channelCount <= 4) {
+    columns = 2; rows = 2;
+  } else if (channelCount <= 9) {
+    columns = 3; rows = 3;
+  } else if (channelCount <= 12) {
+    columns = 4; rows = 3;
+  } else {
+    columns = 4; rows = 4;
+  }
+
+  // Set the grid layout
+  iframeContainer.style.gridTemplateColumns = `repeat(${columns}, 1fr)`;
+  iframeContainer.style.gridTemplateRows = `repeat(${rows}, 1fr)`;
+
+  console.log(`Adjusting grid: ${columns} columns, ${rows} rows for ${channelCount} channels`);
+}
+
+export { adjustIframeSizes };
