+++ templates/index.tmpl
@@ -36,6 +36,6 @@
         {{template "content" .}}
     </div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="/static/scripts.js"></script>
+    <script type="module" src="/static/scripts.js"></script>
 </body>
 </html>
